requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
Django==5.2.11
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-modeltranslation==0.19.19
django-ninja==1.5.3
django-unfold==0.78.1
environs==14.5.0
marshmallow==4.2.2
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
